backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.2.0
pyjwt==2.8.0
pytest==8.0.1
pytest-django==4.8.0
psycopg2==2.9.9